refactor: 좋아요 집계를 product.like_count 에서 like 도메인 소유 product_like_count 테이블로 분리

+++ apps/commerce-api/src/main/java/com/loopers/interfaces/api/product/ProductV1Dto.java
@@ -26,7 +26,7 @@ public class ProductV1Dto {
         String description,
         Long price,
         Integer stock,
-        int likeCount
+        long likeCount
     ) {
         public static ProductResponse from(ProductInfo info) {
             return new ProductResponse(
@@ -49,7 +49,7 @@ public class ProductV1Dto {
         String description,
         Long price,
         Integer stock,
-        int likeCount
+        long likeCount
     ) {
         public static ProductDetailResponse from(ProductDetailInfo info) {
             return new ProductDetailResponse(
